Moved database schema out of db directory to facilitate easier inspection of files at commit.
