+++ .gitignore
@@ -36,3 +36,6 @@ jspm_packages
 # Optional REPL history
 .node_repl_history
 .babel_cache
+
+# Directory for Database 
+db
